requirements.txt
Django==1.9.6
django-filter==0.13.0
django-rest-swagger==0.3.7
djangorestframework==3.3.3
ecdsa==0.13
Fabric==1.11.1
Markdown==2.6.6
mysqlclient==1.3.7
paramiko==1.17.0
Pillow==3.2.0
pycrypto==2.6.1
Pygments==2.1.3
pytz==2016.4
PyYAML==3.11
